Contacts Page Completed
Formatted image and bio section, all that is left is actually writing
out info about myself in the bio section.

Removed one of the cards in the home page, as the "About Me" section
should be in the contacts page and having another one in the home page
is redundant.

Improved the look of the site.